# docs/swagger.yaml
basePath: /api/war-onk
definitions:
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  presenter.GroupRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  presenter.GroupResponse:
    properties:
      active:
        type: boolean
      group_name:
        type: string
      id:
        type: integer
      value:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
host: localhost:8085
info:
  contact:
    email: youremail@example.com
    name: Temi Saputra
  description: Documentation for my REST API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: My REST API
  version: "1.0"
paths:
  /group-create:
    post:
      consumes:
      - application/json
      description: Create a new group
      parameters:
      - description: Group data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presenter.GroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/presenter.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Create a new group
      tags:
      - Group
  /group-delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/presenter.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Delete a group
      tags:
      - Group
  /group-update/{id}:
    put:
      consumes:
      - application/json
      description: Update a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presenter.GroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/presenter.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Update a group
      tags:
      - Group
  /group/{id}:
    get:
      consumes:
      - application/json
      description: Get Group by ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/presenter.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Get Group by ID
      tags:
      - Group
  /groups:
    get:
      consumes:
      - application/json
      description: Get All Group
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Keyword for search
        in: query
        name: keyword
        type: string
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Order type (asc/desc)
        in: query
        name: order_type
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by nick_name
        in: query
        name: nick_name
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/presenter.GroupResponse'
                  type: array
                meta:
                  $ref: '#/definitions/response.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Get All Group
      tags:
      - Group
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"